backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-docx==1.1.0
reportlab==4.4.10
PyMuPDF==1.27.1
pypdf==4.0.1
pdfplumber==0.10.4
spacy==3.7.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.0
email-validator==2.3.0
slowapi==0.1.9
pyspellchecker==0.8.1
pdf2docx==0.5.8
mammoth==1.11.0
selenium==4.15.2
webdriver-manager==4.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
beautifulsoup4==4.12.3
pyjwt==2.8.0
httpx==0.26.0

# Phase 1 Dependencies - Semantic Matching & AI Improvements
sentence-transformers==2.3.1
keybert==0.8.3
language-tool-python==2.7.1
diskcache==5.6.3
